requirements.txt
ultralytics==8.1.0
torch==2.6.0
torchvision==0.21.0
pycocotools>=2.0.7
albumentations>=1.3.1
Pillow>=10.2.0
numpy>=1.26.4
onnxruntime-gpu>=1.20.0
ensemble-boxes>=1.0.9
supervision>=0.18.0
safetensors>=0.4.2
python-dotenv>=1.0
